Main file
This file contains the main class of the TicTacToe game. Functions are used within the class to start the game, make moves and decide the result. There are two players 'X' and 'O'. X is usually the human player while O is usually the computer player.
Run this file to play the game.