Players
This is the file that contains the code of all players: the human player, random computer player, and the minimax computer player (the intelligent one). This file is imported into the Main_game file. This file possesses 4 classes: one for identifying a player and one for each of the three players.